Finish gridpane rework, with notify rather than gui rebuild

+++ Reika/SatisfactoryPlanner/Util/FactoryListener.java
@@ -17,7 +17,7 @@ public interface FactoryListener extends Comparable<FactoryListener> {
 	public void onAddRecipe(Recipe r);
 	public void onRemoveRecipe(Recipe r);
 	public void onSetCount(Recipe r, float count);
-	public void onSetCount(Generator g, float count);
+	public void onSetCount(Generator g, int count);
 	public void onAddProduct(Consumable c);
 	public void onRemoveProduct(Consumable c);
 	public void onAddSupply(ResourceSupply s);
